requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
pandas
openpyxl
flask
gunicorn
flask-cors 
azure-core
azure-ai-formrecognizer